+++ R/summary.nprcmanagErr.R
@@ -83,3 +83,29 @@ summary.nprcmanagErr <- function(object, ...) {
   class(txt) <- "summary.nprcmanagErr"
   txt
 }
+#' @rdname summary
+#' @return object of class summary.nprcmanagGV
+#' @method summary nprcmanagGV
+#' @importFrom stringi stri_c
+#' @export
+summary.nprcmanagGV <- function(object, ...) {
+  gvReport <- object
+  stopifnot(inherits(gvReport, "nprcmanagGV"))
+  rpt <- gvReport[["report"]]
+  kmat <- gvReport[["kinship"]]
+  f <- gvReport[["total"]]
+  mf <- gvReport[["maleFounders"]]
+  ff <- gvReport[["femaleFounders"]]
+  fe <- gvReport[["fe"]]
+  fg <- gvReport[["fg"]]
+  txt <- "The genetic value report"
+  txt <- c(txt, stri_c("Individuals in Pedigree: ", nrow(rpt)))
+  txt <- c(txt, stri_c("Male Founders: ", mf, "\nFemale Founders: ", ff, "\nTotal Founders: ", f))
+  txt <- c(txt, stri_c("Founder Equivalents: ", round(fe, 2)))
+  txt <- c(txt, stri_c("Founder Genome Equivalents: ", round(fg, 2)))
+  txt <- c(txt, stri_c("Live Offspring: ", sum(rpt$livingOffspring)))
+  txt <- c(txt, stri_c("High Value Individuals: ", nrow(rpt[rpt$value == "High Value", ])))
+  txt <- c(txt, stri_c("Low Value Individuals: ", nrow(rpt[rpt$value == "Low Value", ])))
+  class(txt) <- "summary.nprcmanagGV"
+  txt
+}